+++ src/v2/client.v2.write.ts
@@ -1,5 +1,14 @@
 import { API_V2_PREFIX } from '../globals';
 import TwitterApiv2ReadOnly from './client.v2.read';
+import type {
+  StreamingV2AddRulesParams,
+  StreamingV2DeleteRulesParams,
+  StreamingV2UpdateRulesAddResult,
+  StreamingV2UpdateRulesDeleteResult,
+  StreamingV2UpdateRulesParams,
+  StreamingV2UpdateRulesQuery,
+  StreamingV2UpdateRulesResult
+} from '../types';
 
 /**
  * Base Twitter v2 client with read/write rights.
@@ -12,5 +21,24 @@ export default class TwitterApiv2ReadWrite extends TwitterApiv2ReadOnly {
    */
   public get readOnly() {
     return this as TwitterApiv2ReadOnly;
+  }
+
+
+  /* Streaming API */
+
+  public updateStreamRules(options: StreamingV2AddRulesParams, query?: Partial<StreamingV2UpdateRulesQuery>): Promise<StreamingV2UpdateRulesAddResult>;
+  public updateStreamRules(options: StreamingV2DeleteRulesParams, query?: Partial<StreamingV2UpdateRulesQuery>): Promise<StreamingV2UpdateRulesDeleteResult>;
+  /**
+   * Add or delete rules to your stream.
+   * To create one or more rules, submit an add JSON body with an array of rules and operators.
+   * Similarly, to delete one or more rules, submit a delete JSON body with an array of list of existing rule IDs.
+   * https://developer.twitter.com/en/docs/twitter-api/tweets/filtered-stream/api-reference/post-tweets-search-stream-rules
+   */
+  public updateStreamRules(options: StreamingV2UpdateRulesParams, query: Partial<StreamingV2UpdateRulesQuery> = {}) {
+    return this.post<StreamingV2UpdateRulesResult>(
+      'tweets/search/stream/rules',
+      options,
+      { query },
+    );
   }
 }
